Use a Groovy specific package for the Groovy DSL
This will make things easier for when the Kotlin DSL is merged, and
for being compatible with Java modules.
